refactor: replace imports from crypto_prices to currency_class for consistency
