+++ tests/helpers/test_crypto_conversion.py
@@ -6,7 +6,7 @@ import pytest
 from nectar.amount import Amount
 
 from v4vapp_backend_v2.helpers.crypto_conversion import CryptoConversion
-from v4vapp_backend_v2.helpers.crypto_prices import Currency
+from v4vapp_backend_v2.helpers.currency_class import Currency
 
 
 @pytest.fixture(autouse=True)
